feat: create responses agent + testing
