+++ src/core/modules/chatgpt/agents/ChatGPTAgent_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.core.modules.chatgpt.agents.ChatGPTAgent import create_agent
+from src.core.modules.chatgpt.agents.ChatGPTResponsesAgent import create_agent
 
 @pytest.fixture
 def agent():
@@ -13,10 +13,11 @@ def test_agent_name(agent):
     assert "ChatGPT" in result, result
 
 def test_search_news(agent):
-    result = agent.invoke("Which team won the last men's Champions League?")
+    result = agent.invoke("Which team won the last men's football Champions League?")
 
     assert result is not None, result
-    assert "Paris Saint-Germain" in result, result
+    assert "Paris" in result, result
+    assert "Milan" in result, result
 
 def test_search_news_with_datetime(agent):
     from datetime import datetime
@@ -25,3 +26,18 @@ def test_search_news_with_datetime(agent):
 
     assert result is not None, result
     assert datetime.now().strftime("%Y-%m-%d") in result, result
+
+def test_analyze_image(agent):
+    image_url = "https://upload.wikimedia.org/wikipedia/commons/thumb/d/dd/Gfp-wisconsin-madison-the-nature-boardwalk.jpg/2560px-Gfp-wisconsin-madison-the-nature-boardwalk.jpg"
+    result = agent.invoke(f"Analyze the image: {image_url}")
+
+    assert result is not None, result
+    assert "Eiffel Tower" not in result, result
+    assert "boardwalk" in result, result
+
+def test_analyze_pdf(agent):
+    pdf_url = "https://assets.group.accor.com/yrj0orc8tx24/NHhsrTDX0ecCjky5bsvSV/6b631aa9885bf8c3ed84fb323a91e310/ACCOR_IMPACT_REPORT_2023.pdf"
+    result = agent.invoke(f"Extract people cited in the PDF: {pdf_url}")
+
+    assert result is not None, result
+    assert "sébastien bazin" in result.lower(), result
